+++ rusk/src/lib/node/vm.rs
@@ -161,8 +161,13 @@ impl VMExecution for Rusk {
                     anyhow::anyhow!("Cannot check account: {e}")
                 })?;
 
-                let max_value =
-                    tx.value() + tx.deposit() + tx.gas_limit() * tx.gas_price();
+                let max_value = tx
+                    .gas_limit()
+                    .checked_mul(tx.gas_price())
+                    .and_then(|v| v.checked_add(tx.value()))
+                    .and_then(|v| v.checked_add(tx.deposit()))
+                    .ok_or(anyhow::anyhow!("Value spent will overflow"))?;
+
                 if max_value > account_data.balance {
                     return Err(anyhow::anyhow!(
                         "Value spent larger than account holds"